% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agl_Vers_XYZ.R
\name{Agl_Vers_XYZ}
\alias{Agl_Vers_XYZ}
\title{Distances and angles to coordinates}
\usage{
Agl_Vers_XYZ(dist_P2P1, vAngle_XY, vAngle_XZ)
}
\arguments{
\item{dist_P2P1}{Segment length (m)}

\item{vAngle_XY}{Angle between the segment and the XY plane (radian)}

\item{vAngle_XZ}{Angle between the segment and the XZ plane (radian)}
}
\value{

}
\description{
Transform distances and angles into point coordinates
}
