% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_bend.R
\name{compute_error}
\alias{compute_error}
\title{Compute error}
\usage{
compute_error(
  x,
  y,
  field_data,
  unbent_data,
  type = c("all", "bending", "torsion"),
  ...
)
}
\arguments{
\item{x, y}{Either the elasticity modulus (bending, MPa) or Shear modulus (torsion, MPa) (see dtails)}

\item{field_data}{Field data (list of data.frames with x, y and z coordinates of the points)}

\item{unbent_data}{List of outputs from \code{unbend(field_data)} + \code{\link[=distance_weight_sine]{distance_weight_sine()}}}

\item{type}{Type of optimization (either "bending","torsion" or "all")}

\item{...}{Further parameters to pass to \code{\link[=bend]{bend()}}}
}
\value{
The quadratic error of simulation
}
\description{
Compute the error of simulation usin field data, field data that is
un-bent, and bending parameters. Mainly used from \code{\link[=optimize_bend]{optimize_bend()}}.
}
\details{
x an y depends on the optimization type. \code{x} is the elasticity modulus if "bending",
or shear modulus if "torsion". The \code{y} is the contrary. This is due to
the fact that \code{\link[stats:optimize]{stats::optimize()}} always use the first argument for optimization.
If \code{type == "all"}, then \code{x} is the elasticity modulus.
}
\keyword{internal}
