% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeformFlexTor.R
\name{DeformFlexTor}
\alias{DeformFlexTor}
\title{Deform}
\usage{
DeformFlexTor(matPoints, pas, Ncalc, Nboucle, verbose = TRUE)
}
\arguments{
\item{matPoints}{Point matrix}

\item{pas}{Length of the segments that discretize the object (m).}

\item{Ncalc}{Number of points used in the grid discretizing the section.}

\item{Nboucle}{Number of iterations to compute the torsion and bending}

\item{verbose}{Boolean. Return information during procress?}
}
\value{
A matrix: \link{PtsX; PtsY; PtsZ; PtsDist; PtsAglXY; PtsAglXZ; PtsAglTor}
}
\description{
Compute the deformation by applying both bending and torsion
}
\examples{
library(data.table)
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "deformation")
matExp = data.table::fread(filepath, data.table = FALSE)
matPoints = CreaPoints(2000,400,matExp)
# matPoints is the matrix of points for a straight rachis. So it reads
# measured information that is already bent, and put it back to a straight line.
pas= 0.02 # in meter. -> Length of the segments that discretize the object.
Ncalc= 100 # number of points used in the grid that discretized the section.
Nboucle= 15 # if we want to compute the torsion after the bending step by step instead of
# all bending and then torsion.
matRes = DeformFlexTor(matPoints, pas, Ncalc, Nboucle, TRUE); # Computes the deformation -> this is the actual model
}
