% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_distance.R
\name{distance_weight_sine}
\alias{distance_weight_sine}
\title{Weight distance of application}
\usage{
distance_weight_sine(x, dF = 0.1)
}
\arguments{
\item{x}{X position of the segments}

\item{dF}{Amplitude (m)}
}
\value{
The distance at which the left/right weights are applied from the beam axis. Both
sides are considered equal.
}
\description{
Computes the distance of application of the right/left weights using a sine
evolution along the beam.
}
\examples{
distance_weight_sine(c(1,2,3))
}
