% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpPoints.R
\name{InterpPoints}
\alias{InterpPoints}
\title{Point interpolation}
\usage{
InterpPoints(matPoints, pas)
}
\arguments{
\item{matPoints}{Point matrix (from \code{\link[=CreaPoints]{CreaPoints()}})}

\item{pas}{Distance needed between interpolated points (m)}
}
\value{
A list.
}
\description{
Point interpolation
}
