% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bent.R
\name{plot_bending}
\alias{plot_bending}
\alias{plot_bending_3d}
\title{Plot bent}
\usage{
plot_bending(...)

plot_bending_3d(...)
}
\arguments{
\item{...}{A list of X, Y, Z data.frames for each beam of beam state}
}
\value{
A ggplot or a plotly object (see \code{\link[plotly:plot_ly]{plotly::plot_ly()}})
}
\description{
Plot bent
}
\examples{
file_path = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
field_data = read_mat(file_path)
# Un-bending the field measurements:
df_unbent = unbend(field_data)

# Adding the distance of application of the left and right weight:
df_unbent$distance_application = distance_weight_sine(df_unbent$x)

# (Re-)computing the deformation:
df_bent = bend(df_unbent, elastic_modulus = 2000, shear_modulus = 400, step = 0.02,
               points = 100, iterations = 15, verbose = TRUE)

# And finally, plotting:
plot_bending(bent = df_bent, "un-bent" = df_unbent)

# And in 3d:
# plot_bending_3d(bent = df_bent, "un-bent" = df_unbent)
}
