% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bent.R
\name{plot_bent}
\alias{plot_bent}
\alias{plot_bent_3d}
\title{Plot bent}
\usage{
plot_bent(bent, unbent = NULL, ...)

plot_bent_3d(bent, unbent = NULL, ...)
}
\arguments{
\item{bent}{Bent data.frame}

\item{unbent}{Initial state data.frame (as from \code{\link[=unbend]{unbend()}})}

\item{...}{Further parameters.}
}
\value{
A plotly object, see \code{\link[plotly:plot_ly]{plotly::plot_ly()}}
}
\description{
Plot bent
}
\examples{
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
# Un-bending the field measurements:
df = unbend(2000,400, read_mat(filepath))
df_bend = bend(df, step = 0.02, points = 100, iterations = 15, verbose = TRUE)
plot_bent(df_bend,df)
}
