% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unbent.R
\name{plot_unbent}
\alias{plot_unbent}
\title{Plot unbent}
\usage{
plot_unbent(unbent, meas = NULL)
}
\arguments{
\item{unbent}{Unbent data.frame}

\item{meas}{Field measurement data.frame (as from \code{\link[=read_mat]{read_mat()}})}

\item{...}{Further parameters.}
}
\value{
A plotly object, see \code{\link[plotly:plot_ly]{plotly::plot_ly()}}
}
\description{
Plot unbent
}
\examples{
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
meas = read_mat(filepath)
df_unbend = unbend(MOE = 2000, CIS = 400, df = meas)
plot_unbent(df_unbend,meas)
}
