% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat.R
\name{read_mat}
\alias{read_mat}
\title{Read field matrix}
\usage{
read_mat(path)
}
\arguments{
\item{path}{Path to the file}
}
\value{
A formatted \code{\link[=data.frame]{data.frame()}}
}
\description{
Read the matrix from field measurements.
}
\examples{
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "deformation")
matExp = read_mat(filepath)
}
