% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InertieFlexRota.R
\name{Remplir}
\alias{Remplir}
\title{Fill in the matrix}
\usage{
Remplir(section, sct)
}
\arguments{
\item{section}{section matrix}

\item{sct}{Section type (see details)}
}
\value{
The matrix
}
\description{
Compute the matrix according to the section shape
}
\examples{
Remplir(matrix(0, nrow = 6, ncol = 10), 1)
}
