% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rota_Inverse_YZ.R
\name{Rota_Inverse_YZ}
\alias{Rota_Inverse_YZ}
\title{Inverse rotation}
\usage{
Rota_Inverse_YZ(OP, Agl_Y, Agl_Z)
}
\arguments{
\item{OP}{x, y, z coordinates}

\item{Agl_Y}{Rotation angle around Y (radian)}

\item{Agl_Z}{Rotation angle around Z (radian)}
}
\value{
The rotated point
}
\description{
Rotate around the Z axis and then the Y axis
}
\examples{
Rota_Inverse_YZ(matrix(rep(0,3), ncol = 1), 0.8517207, 0)
}
