% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rota_YZ.R
\name{Rota_YZ}
\alias{Rota_YZ}
\title{Rotate}
\usage{
Rota_YZ(OP, Agl_Y, Agl_Z)
}
\arguments{
\item{OP}{A 3d point (x,y,z)}

\item{Agl_Y}{Rotation angle around Y (radian)}

\item{Agl_Z}{Rotation angle around Z (radian)}
}
\value{
The rotated point
}
\description{
Rotate point around Y and (then) Z axes.
}
\examples{
Rota_YZ(c(2.681,0,0),0.85172,0)
}
