% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbend.R
\name{unbend}
\alias{unbend}
\title{Unbend}
\usage{
unbend(MOE, CIS, df)
}
\arguments{
\item{MOE}{Elastic modulus}

\item{CIS}{Shear modulus}

\item{df}{Experimental data (usually read using \code{\link[=read_mat]{read_mat()}}, see details)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}:
\itemize{
\item x: x coordinate
\item y: y coordinate
\item z: z coordinate
\item type: section type. 1: triangle (bottom-oriented); 2 : rectangle; 3 : triangle (top-oriented);
4 : ellipsis; 5 : circle.
\item width (m): section width
\item height (m): section height
\item mass (kg): mass of the section at the point
\item mass_right (kg): mass carried by the object, on the right side
\item mass_left (kg): mass carried by the object, on the left side
\item elastic_modulus (MPa): elasticity modulus (bending)
\item shear_modulus (MPa): shear modulus (torsion)
\item inclination (degree): insertion angle of the first section (only first value is used)
\item distance_application (m): application distances for the left and right weights
}
}
\description{
Takes measurements from field that are already bent (torsion + bending) and transform them
to a straight line (un-bending).
Mainly used to compute the matrix for input to \code{\link[=bend]{bend()}} from experimental points.
}
\details{
df should be a formatted \code{\link[=data.frame]{data.frame()}}, with each row being a point and each columns being:
\itemize{
\item distance (m): distance between the previous point and this point (first value should be positive)
\item type: section type. 1: triangle (bottom-oriented); 2 : rectangle; 3 : triangle (top-oriented);
4 : ellipsis; 5 : circle.
\item width (m): section width
\item height (m): section height
\item inclination (degree): insertion angle of the first point (only first value is used)
\item torsion (degree): torsion angle
\item x: x coordinate (optional, only for plotting)
\item y: y coordinate (optional, only for plotting)
\item z: z coordinate (optional, only for plotting)
\item mass (kg): mass of the section at the point
\item mass_right (kg): mass carried by the object, on the right side
\item mass_left (kg): mass carried by the object, on the left side
}
}
\examples{
\dontrun{
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
df = read_mat(filepath)
unbend(MOE = 2000, CIS = 400, df = df)
}
}
