% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XYZ_Vers_Agl.R
\name{XYZ_Vers_Agl}
\alias{XYZ_Vers_Agl}
\title{Title}
\usage{
XYZ_Vers_Agl(vecX, vecY, vecZ)
}
\arguments{
\item{vecX}{X coordinates (m)}

\item{vecY}{Y coordinates (m)}

\item{vecZ}{Z coordinates (m)}
}
\value{
A list:
\itemize{
\item dist_P2P1: Length of the segment;
\item vAngle_XY: Angle between the segment and the XY plane (radian);
\item vAngle_XZ: Angle between the segment and the XZ plane (radian)
}
}
\description{
Title
}
\examples{
library(data.table)
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "deformation")
df = read_mat(filepath)
df = unbend(2000,400,df)
XYZ_Vers_Agl(df[1,],df[2,],df[3,])

}
