% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreaPoints.R
\name{CreaPoints}
\alias{CreaPoints}
\title{Experimental matrix}
\usage{
CreaPoints(MOE, CIS, matExp)
}
\arguments{
\item{MOE}{Elastic modulus}

\item{CIS}{Shear modulus}

\item{matExp}{Experimental matrix}
}
\value{
The point matrix
}
\description{
Build the matrix from the experimental points
}
\note{
Lengths in meters, angles in degrees and mass in kg in matExp.
}
\examples{
\dontrun{
library(data.table)
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "deformation")
matExp = data.table::fread(filepath, data.table = FALSE)
CreaPoints(2000,400,matExp)
}
}
