% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbend.R
\name{unbend}
\alias{unbend}
\title{Unbend}
\usage{
unbend(df)
}
\arguments{
\item{df}{Experimental data (usually read using \code{\link[=read_mat]{read_mat()}}, see details)}
}
\value{
The input \code{\link[=data.frame]{data.frame()}} with modified (or enriched with):
\itemize{
\item x: x coordinate
\item y: y coordinate
\item z: z coordinate
}
}
\description{
Remove torsion and bending of a bent beam, i.e. transform it
into a straight line, while keeping its insertion angle (inclination angle of the first segment).
}
\details{
df should be a formatted \code{\link[=data.frame]{data.frame()}}, with each row being a point and each columns being:
\itemize{
\item distance (m): distance between the previous point and this point (first value should be positive)
\item inclination (degree): insertion angle of the first point (only first value is used)
}
}
\note{
Mainly used to compute the matrix for input to \code{\link[=bend]{bend()}} from experimental points.
}
\examples{
\dontrun{
filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
df = read_mat(filepath)
unbend(df = df)
}
}
